Updating jwt.Attempt() to accept custom payload (#39)
* fix(jwt): Allowed jwt.attempt() to accept an optional payload

With recent work on jwt.generate(), that function signature was updated to allow a user to push a
custom payload into the JWT that was created. Since jwt.attempt() uses jwt.generate() internally,
this fix just updates the jwt.attempt() method to allow also adding a custom payload

adonisjs/adonis-auth/issues/38

* test(jwt): JWT.attempt() test for updated method signature

Added a test for verifying JWT.attempt() works properly when you include a custom payload for the
token generation.