refactor(auth): use uuidv4 over v1
The generate modified in this change seems to be what is generating API (and session?) tokens. If this is the case, then I believe that using the v4() method instead of v1() would be more desirable as it would provide higher entropy and so be harder to reproduce. v4() is completely random while v1() is time-based.